refactor: history 라우터 분리
