+++ apps/backend/api/main.py
@@ -82,7 +82,7 @@ from src.db import close_pool, init_pool
 from src.dependencies import warmup_dependencies
 from rag_pipeline.observability import initialize_langfuse, shutdown_langfuse
 from rag_pipeline.retrieval.common import EmbeddingError, LLMError, LLMTimeoutError, RetrievalError
-from src.routers import health, qa, graph
+from src.routers import health, qa, graph, history
 
 
 # ─── 앱 생명주기 관리 (Lifespan) ──────────────────────────────────────────────
@@ -183,4 +183,5 @@ async def internal_exception_handler(request: Request, exc: Exception):
 
 app.include_router(health.router)
 app.include_router(qa.router, prefix="/api/v1/qa")
+app.include_router(history.router, prefix="/api/v1/qa")
 app.include_router(graph.router, prefix="/api/v1/graph")
